# Conflicts:
#	build/.cmake/api/v1/reply/cache-v2-245f4b08d08ed2c7239a.json
#	build/.cmake/api/v1/reply/directory-_deps.ftxui-build-Debug-92547a97cc005e646df9.json
#	build/.cmake/api/v1/reply/index-2025-04-10T02-19-47-0705.json
#	build/.cmake/api/v1/reply/target-UI-Debug-58d7ee688d81bc8a5749.json
#	build/.cmake/api/v1/reply/target-component-Debug-492350a336aaac86c5ae.json
#	build/.cmake/api/v1/reply/target-dom-Debug-2a94e7f5370ecc74a248.json
#	build/.cmake/api/v1/reply/target-screen-Debug-703547bfed55189d4897.json
#	build/.cmake/api/v1/reply/toolchains-v1-ee793ea44d57a62b2395.json
#	build/CMakeCache.txt
#	build/CMakeFiles/3.28.3/CMakeCCompiler.cmake
#	build/CMakeFiles/3.28.3/CMakeCXXCompiler.cmake
#	build/CMakeFiles/3.28.3/CMakeDetermineCompilerABI_C.bin
#	build/CMakeFiles/3.28.3/CMakeDetermineCompilerABI_CXX.bin
#	build/CMakeFiles/3.28.3/CompilerIdC/CMakeCCompilerId.c
#	build/CMakeFiles/3.28.3/CompilerIdC/a.out
#	build/CMakeFiles/3.28.3/CompilerIdCXX/CMakeCXXCompilerId.cpp
#	build/CMakeFiles/3.28.3/CompilerIdCXX/a.out
#	build/CMakeFiles/CMakeConfigureLog.yaml
#	build/CMakeFiles/CMakeDirectoryInformation.cmake
#	build/CMakeFiles/CMakeOutput.log
#	build/CMakeFiles/Makefile.cmake
#	build/CMakeFiles/Makefile2
#	build/CMakeFiles/UI.dir/build.make
#	build/CMakeFiles/UI.dir/flags.make
#	build/CMakeFiles/UI.dir/link.txt
#	build/CMakeFiles/UI.dir/src/Display.cpp.o
#	build/CMakeFiles/clion-Debug-WSL-log.txt
#	build/CMakeFiles/clion-environment.txt
#	build/CPackConfig.cmake
#	build/CPackSourceConfig.cmake
#	build/Catch2/CMakeFiles/CMakeDirectoryInformation.cmake
#	build/Catch2/CMakeFiles/progress.marks
#	build/Catch2/CTestTestfile.cmake
#	build/Catch2/Makefile
#	build/Catch2/cmake_install.cmake
#	build/Catch2/src/CMakeFiles/CMakeDirectoryInformation.cmake
#	build/Catch2/src/CMakeFiles/Catch2.dir/DependInfo.cmake
#	build/Catch2/src/CMakeFiles/Catch2.dir/build.make
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/benchmark/catch_chronometer.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/benchmark/catch_chronometer.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/benchmark/detail/catch_analyse.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/benchmark/detail/catch_analyse.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/benchmark/detail/catch_benchmark_function.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/benchmark/detail/catch_benchmark_function.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/benchmark/detail/catch_run_for_at_least.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/benchmark/detail/catch_run_for_at_least.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/benchmark/detail/catch_stats.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/benchmark/detail/catch_stats.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/generators/catch_generator_exception.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/generators/catch_generator_exception.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/generators/catch_generators.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/generators/catch_generators.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/generators/catch_generators_random.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/generators/catch_generators_random.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_automake.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_automake.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_common_base.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_common_base.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_compact.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_compact.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_console.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_console.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_cumulative_base.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_cumulative_base.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_event_listener.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_event_listener.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_helpers.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_helpers.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_json.cpp.o
#	build/Catch2/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_json.cpp.o.d
#	build/Catch2/src/CMakeFiles/Catch2.dir/cmake_clean.cmake
#	build/Catch2/src/CMakeFiles/Catch2.dir/cmake_clean_target.cmake
#	build/Catch2/src/CMakeFiles/Catch2.dir/compiler_depend.make
#	build/Catch2/src/CMakeFiles/Catch2.dir/flags.make
#	build/Catch2/src/CMakeFiles/Catch2.dir/link.txt
#	build/Catch2/src/CMakeFiles/Catch2.dir/progress.make
#	build/Catch2/src/CMakeFiles/Catch2WithMain.dir/DependInfo.cmake
#	build/Catch2/src/CMakeFiles/Catch2WithMain.dir/build.make
#	build/Catch2/src/CMakeFiles/Catch2WithMain.dir/cmake_clean.cmake
#	build/Catch2/src/CMakeFiles/Catch2WithMain.dir/cmake_clean_target.cmake
#	build/Catch2/src/CMakeFiles/Catch2WithMain.dir/compiler_depend.make
#	build/Catch2/src/CMakeFiles/Catch2WithMain.dir/flags.make
#	build/Catch2/src/CMakeFiles/Catch2WithMain.dir/link.txt
#	build/Catch2/src/CMakeFiles/Catch2WithMain.dir/progress.make
#	build/Catch2/src/CMakeFiles/progress.marks
#	build/Catch2/src/CTestTestfile.cmake
#	build/Catch2/src/Makefile
#	build/Catch2/src/cmake_install.cmake
#	build/Makefile
#	build/Testing/Temporary/LastTest.log
#	build/UI
#	build/_deps/ftxui-build/CMakeFiles/CMakeDirectoryInformation.cmake
#	build/_deps/ftxui-build/CMakeFiles/Export/418a5211b54bdfa9fa029e8a3112c4d3/ftxui-targets.cmake
#	build/_deps/ftxui-build/CMakeFiles/component.dir/build.make
#	build/_deps/ftxui-build/CMakeFiles/component.dir/flags.make
#	build/_deps/ftxui-build/CMakeFiles/dom.dir/build.make
#	build/_deps/ftxui-build/CMakeFiles/dom.dir/flags.make
#	build/_deps/ftxui-build/CMakeFiles/screen.dir/build.make
#	build/_deps/ftxui-build/CMakeFiles/screen.dir/flags.make
#	build/_deps/ftxui-build/Makefile
#	build/_deps/ftxui-build/cmake_install.cmake
#	build/_deps/ftxui-build/doc/CMakeFiles/CMakeDirectoryInformation.cmake
#	build/_deps/ftxui-build/doc/Makefile
#	build/_deps/ftxui-build/examples/CMakeFiles/CMakeDirectoryInformation.cmake
#	build/_deps/ftxui-build/examples/Makefile
#	build/_deps/ftxui-build/ftxui-targets.cmake
#	build/_deps/ftxui-subbuild/CMakeCache.txt
#	build/_deps/ftxui-subbuild/CMakeFiles/CMakeConfigureLog.yaml
#	build/_deps/ftxui-subbuild/CMakeFiles/CMakeDirectoryInformation.cmake
#	build/_deps/ftxui-subbuild/CMakeFiles/CMakeRuleHashes.txt
#	build/_deps/ftxui-subbuild/CMakeFiles/Makefile.cmake
#	build/_deps/ftxui-subbuild/CMakeFiles/Makefile2
#	build/_deps/ftxui-subbuild/CMakeFiles/ftxui-populate.dir/build.make
#	build/_deps/ftxui-subbuild/CMakeLists.txt
#	build/_deps/ftxui-subbuild/Makefile
#	build/_deps/ftxui-subbuild/cmake_install.cmake
#	build/_deps/ftxui-subbuild/ftxui-populate-prefix/src/ftxui-populate-stamp/ftxui-populate-gitclone-lastrun.txt
#	build/_deps/ftxui-subbuild/ftxui-populate-prefix/src/ftxui-populate-stamp/ftxui-populate-gitinfo.txt
#	build/_deps/ftxui-subbuild/ftxui-populate-prefix/src/ftxui-populate-stamp/ftxui-populate-update-info.txt
#	build/_deps/ftxui-subbuild/ftxui-populate-prefix/tmp/ftxui-populate-gitclone.cmake
#	build/_deps/ftxui-subbuild/ftxui-populate-prefix/tmp/ftxui-populate-gitupdate.cmake
#	build/_deps/ftxui-subbuild/ftxui-populate-prefix/tmp/ftxui-populate-mkdirs.cmake
#	build/cmake_install.cmake
#	build/compile_commands.json

+++ src/Elevator.cpp
@@ -452,9 +452,14 @@ void eWaitingForInput::handle(ElevatorSubsystem* context) {
 
         while (context->myElevator.getQueue().empty()) context->cv.wait(lock);
         std::cout << "Acquired Mutex" << std::endl;
+        std::cout << "User Direction: " << context->myElevator.getDirection() << std::endl;
         if (context->myElevator.getDirection() == BROKEN) {
-          context->setState(new BrokenState());
-          context->handle();
+            std::cout << "Entered If: "<< std::endl;
+            context->setState(new BrokenState());
+
+        }
+        else {
+            context->setState(new ProcessRequest());
         }
         context->myElevator.printQueue();
         //review this line. was confused about it before.
@@ -466,7 +471,7 @@ void eWaitingForInput::handle(ElevatorSubsystem* context) {
 
     }
     std::cout << "Elevator " << context->programID << ": Received Request" << std::endl;
-    context->setState(new ProcessRequest());
+
     context->handle();
 }
 
@@ -536,46 +541,45 @@ void Stopped::handle(ElevatorSubsystem* context) {
     std::cout << "Elevator " << context->programID << ": Stopped." << std::endl;
     std::this_thread::sleep_for(std::chrono::seconds(1));
     // remove the floor we just arrived at
+    if (!context->stateTest) {
+        if (true) {
+            std::unique_lock<std::mutex> lock(context->mtx);
+            int allowed = context->deleteEntry(context->myElevator.getQueue(), context->myElevator.getCurrentFloor());
+            auto direction = context->myElevator.getDirection();
+            context->myElevator.setDirection(IDLE);
+            if (context->myElevator.user_direction.empty() && context->myElevator.getQueue().empty()) {
+                std::cout << "Elevator " << context->programID << ": Done servicing queue" << std::endl;
+                std::cout << "Elevator " << context->programID << ": Current Capacity is: " << context->capacity << std::endl;
+            } else {
+                std::cout << "adding to user queue."<< std::endl;
+                // add all the user button requets into the queue
 
-    if (true) {
-        std::unique_lock<std::mutex> lock(context->mtx);
-        int allowed = context->deleteEntry(context->myElevator.getQueue(), context->myElevator.getCurrentFloor());
-        auto direction = context->myElevator.getDirection();
-        context->myElevator.setDirection(IDLE);
-        if (context->myElevator.user_direction.empty() && context->myElevator.getQueue().empty()) {
-            std::cout << "Elevator " << context->programID << ": Done servicing queue" << std::endl;
-            std::cout << "Elevator " << context->programID << ": Current Capacity is: " << context->capacity << std::endl;
-        } else {
-            std::cout << "adding to user queue."<< std::endl;
-            // add all the user button requets into the queue
-
-            if (allowed > 0) { // If we were picking up passengers
-                context->userQueuePop(context->myElevator.getCurrentFloor(), context->myElevator.user_direction, allowed);
-            }
-            else {
-                if (context->front(context->myElevator.getQueue()).first == context->myElevator.getCurrentFloor()) {
-                    context->myElevator.setDirection(direction);
-                }else {
-                    context->calcdirection(context->front(context->myElevator.getQueue()).first);
+                if (allowed > 0) { // If we were picking up passengers
+                    context->userQueuePop(context->myElevator.getCurrentFloor(), context->myElevator.user_direction, allowed);
                 }
-                std::cout << "Current Direction (stopped state): " << context->myElevator.getDirection() << std::endl;
-                context->sortMapInPlace(context->myElevator.getQueue(), context->myElevator.getDirection() == UP);
+                else {
+                    if (context->front(context->myElevator.getQueue()).first == context->myElevator.getCurrentFloor()) {
+                        context->myElevator.setDirection(direction);
+                    }else {
+                        context->calcdirection(context->front(context->myElevator.getQueue()).first);
+                    }
+                    std::cout << "Current Direction (stopped state): " << context->myElevator.getDirection() << std::endl;
+                    context->sortMapInPlace(context->myElevator.getQueue(), context->myElevator.getDirection() == UP);
+                }
+                auto pair = context->front(context->myElevator.getQueue());
+                std::cout << "used front."<< std::endl;
+                if (!pair.second.first.empty()) { // If we are picking up passengers
+                    context->myElevator.user_direction = pair.second.first;
+                }
+                else { //dropping off passengers
+                    context->myElevator.user_direction = "";
+                }
+                context->myElevator.floor_to_go_to = pair.first;
+                context->myElevator.printQueue();
+                std::cout << "Elevator " << context->programID << ": Current Capacity is: " << context->capacity << std::endl;
             }
-            auto pair = context->front(context->myElevator.getQueue());
-            std::cout << "used front."<< std::endl;
-            if (!pair.second.first.empty()) { // If we are picking up passengers
-                context->myElevator.user_direction = pair.second.first;
-            }
-            else { //dropping off passengers
-                context->myElevator.user_direction = "";
-            }
-            context->myElevator.floor_to_go_to = pair.first;
-            context->myElevator.printQueue();
-            std::cout << "Elevator " << context->programID << ": Current Capacity is: " << context->capacity << std::endl;
         }
     }
-
-
     if (context->doorsJammed) {
       context->setState(new JammedState());
     } else {
@@ -632,17 +636,22 @@ void DoorsClosed::handle(ElevatorSubsystem* context) {
 }
 
 void BrokenState::handle(ElevatorSubsystem* context) {
-    std::cout << "Elevator " << context->programID << ": Stuck Between Floors." << std::endl; // <--- START HERE FOR NEXT ITERATION
-    while (1){}
+    if (!context->stateTest) {
+        std::cout << "Elevator " << context->programID << ": Stuck Between Floors." << std::endl; // <--- START HERE FOR NEXT ITERATION
+        while (1){}
+    }
 }
 
 void JammedState::handle(ElevatorSubsystem* context) {
+
     std::cout << "Doors Jammed.... Initated Fixed Doors Routine." << std::endl;
     std::this_thread::sleep_for(std::chrono::seconds(5));
     context->doorsJammed = false;
     std::cout << "Doors have been fixed!!" << std::endl;
-    context->setState(new DoorsOpened());
-    context->handle();
+    if (!context->stateTest) {
+        context->setState(new DoorsOpened());
+        context->handle();
+    }
 }
 
 //
